compat: backport to FreeBSD 12.2
This should allow us to get more testing coverage earlier.

This port here is also a bit janky. I really don't like the taskqgroup
business, having to copy and paste those structs. And this isn't well
tested, either. But, it's a start.

This distinguishes between compat.h and support.h, though both header
files are intended to operate in more or less the same way.  It's
important to keep some discipline between things that we're backporting
and things that aren't _yet_ upstream or are shims for OpenBSD.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>