wg_noise: ensure non-zero'd handshakes have a valid local index
As reported by: https://marc.info/?l=openbsd-bugs&m=161618496905444&w=2

In particular, when consuming an initiation, we don't generate the
index until creating the response (which is incorrect). If we attempt
to create an initiation between these processes, we drop any
outstanding handshake which in this case has index 0 as set when
consuming the initiation.

The fix attached is to generate the index when consuming the initiation
so that any spurious initiation creation can drop a valid index. The
patch also consolidates setting fields on the handshake.

Signed-off-by: Matt Dunwoodie <ncon@noconroy.net>