crypto: chacha and poly in same loop
This is a fixup of f685f466, where previously we chacha'd in a
different loop to poly'ing. Now we do in the same loop to keep the cache
hot. In practice this didn't result in an (easily) observable change,
which could be due to only having 1-2 mbufs in a chain. However this is
still the preferred way to do it.

Signed-off-by: Matt Dunwoodie <ncon@noconroy.net>