if_wg: better loop detection
While it was nice to have per peer loop detection, it was not meant to
be. The loop tag has a tag type == 0, which conflicts with other tags.
Therefore we want to at least be a little bit more sure that the tag
cookie is unique to the loop tag. I guess the peer address was also
quite hacky so on the other side, I'm glad to be rid of that.

Now we have a loop of 8 (to any peer) which should be good enough for an
edge case operation.

Signed-off-by: Matt Dunwoodie <ncon@noconroy.net>