if_wg: remove M_WAITOK, check return codes on init
Here we remove all M_WAITOK checks, because we don't want to hang while
trying to allocate memory. It is better to return an error so the user
can try again later.

We also make sure to check all the return codes in peer and interface
allocation. The structure of those functions is:

1) Allocate all memory
2) Initialise fields in order of the struct
3) Cleanup gotos

Signed-off-by: Matt Dunwoodie <ncon@noconroy.net>