if_wg: handle if_transmit and if_output properly
The netmap code goes directly to if_transmit, which means it'll bypass
if_output, in which case, there's no packet allocated. Also, we're
relying on if_output's sockaddr structure to be legit, but who knows
what types of userspace hijynxes can forge this. Rather than relying on
that kind of black magic, determine the AF from the actual packet
contents. But still insist that it agrees with the sockaddr.

The extraction of the type from AF_UNSPEC follows the same pattern as
if_gif and if_gre.

We also use this as an opportunity to send icmp error messages in the
right place.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>