if_wg: defragment mbufs early on
This makes the crypto a *lot* faster. We might later revert this if we
use opencrypto's fancy page table mapping scheme. But for now, it's
useful. We do it early, rather than before calling decrypt/encrypt, so
that the various other m_pullups that we have wind up being no-ops.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>