if_wg: return to m temporary variable style
The rest of the code uses this, so go with it for now. Maybe later ncon
will want to clean up everything to be this way, but for now keep it
consistent.

This partially reverts commit a1fdf6646b16ec26c741089102346f5455dc5fed,
but doesn't reintroduce the bug that it had fixed.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>