if_wg: avoid scheduling excessive tasks for encryption/decryption
Previously the wg_encrypt/decrypt_dispatch routines would wakeup all
of the non-scheduled tasks.  This resulted in waking up several tasks
which did no useful work but did add overhead in terms of extra
context switches, etc.  In a single-threaded iperf3 TCP test on an
8-CPU system, only ~13% of the scheduled tasks did any work.

Instead, adopt a similar scheme to that used in the Linux wireguard
driver in which each call only wakes up a single task and uses
per-queue CPU iterators to round-robin among CPUs when choosing a task
to schedule.  This improved throughput in the iperf3 TCP test by more
than double and ~65% of scheduled tasks did useful work.

Signed-off-by: John Baldwin <jhb@FreeBSD.org>