if_wg: account for input function returning a boolean
Since 742e7210 ("udp: allow udp_tun_func_t() to indicate it did not eat
the packet"), wg_input must return a boolean. We force a cast for old
kernels. It'd be nicer to work around this in compat.h, but we can't
because FreeBSD's headers have dependencies we can't resolve from there.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>